+++ protected/humhub/modules/content/components/ContentTagActiveQuery.php
@@ -44,7 +44,7 @@ class ContentTagActiveQuery extends ActiveQuery
         ];
 
         if ($user !== null) {
-            if (!$user->canViewAllContent(Space::class)) {
+            if (!$user->canManageContent()) {
                 // User must be a space's member OR a space is not private
                 $this->leftJoin('space_membership AS rMembership', 'rMembership.space_id = rSpace.id AND rMembership.user_id = :userId', [':userId' => $user->id]);
                 $conditions['space'][] = [
@@ -52,9 +52,6 @@ class ContentTagActiveQuery extends ActiveQuery
                     ['rMembership.status' => Membership::STATUS_MEMBER],
                     ['!=', 'rSpace.visibility', Space::VISIBILITY_NONE],
                 ];
-            }
-
-            if (!$user->canViewAllContent(User::class)) {
                 // User can view only content of own profile
                 $conditions['user'][] = ['rUser.id' => $user->id];
             }