+++ protected/humhub/modules/content/permissions/ManageContent.php
@@ -9,8 +9,10 @@
 namespace humhub\modules\content\permissions;
 
 use humhub\libs\BasePermission;
-use humhub\modules\user\models\User;
+use humhub\modules\content\Module;
 use humhub\modules\space\models\Space;
+use humhub\modules\user\models\Group;
+use humhub\modules\user\models\User;
 use Yii;
 
 /**
@@ -25,6 +27,8 @@ class ManageContent extends BasePermission
      * @inheritdoc
      */
     protected $moduleId = 'content';
+
+    protected $defaultState = self::STATE_DENY;
 
     /**
      * @inheritdoc
@@ -62,6 +66,43 @@ class ManageContent extends BasePermission
      */
     public function getDescription()
     {
-        return Yii::t('CommentModule.permissions', 'Can manage (e.g. archive, stick, move or delete) arbitrary content');
+        return $this->contentContainer ?
+            Yii::t('CommentModule.permissions', 'Can manage (e.g. archive, stick, move or delete) arbitrary content') :
+            Yii::t('CommentModule.permissions', 'Can manage (e.g. edit or delete) all content (even private)');
+    }
+
+    /**
+     * @inheritdoc
+     */
+    public function init()
+    {
+        if (!$this->contentContainer) {
+            /** @var Module $module */
+            $module = Yii::$app->getModule('content');
+            if ($module->enableGlobalManageContentPermission) {
+                $this->fixedGroups[] = Group::getAdminGroupId();
+            }
+        }
+
+        parent::init();
+    }
+
+    /**
+     * {@inheritdoc}
+     *
+     * Note: that this function always returns state self::STATE_ALLOW for the administration
+     * group, this behaviour can't be overwritten by means of the configuration.
+     */
+    public function getDefaultState($groupId)
+    {
+        if (!$this->contentContainer && $groupId === Group::getAdminGroupId()) {
+            /** @var Module $module */
+            $module = Yii::$app->getModule('content');
+            if ($module->enableGlobalManageContentPermission) {
+                return self::STATE_ALLOW;
+            }
+        }
+
+        return parent::getDefaultState($groupId);
     }
 }